backend/requirements.txt
azure-functions
azure-storage-blob
pandas
numpy
azure-cosmos